+++ src/test/java/com/powsybl/openloadflow/ac/DistributedSlackOnGenerationTest.java
@@ -31,7 +31,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public class DistributedGenerationSlackTest {
+public class DistributedSlackOnGenerationTest {
 
     private Network network;
     private Generator g1;
